+++ src/pages/minsuKim/comment/CommentMinsu.scss
@@ -3,74 +3,16 @@
 $backgroundColor: rgb(250, 250, 250);
 $border: 1px solid rgba(0, 0, 0, 0.2);
 
-.commentWrapper {
-  padding: 15px 20px;
-
-  .commentIcon {
-    display: flex;
-    justify-content: space-between;
-    font-size: 25px;
-
-    div {
-      display: flex;
-      width: 15%;
-      justify-content: space-between;
-    }
-
-    .fa-heart {
-      color: tomato;
-    }
-
-    .fa-comment {
-      color: blue;
-    }
-
-    .fa-envelope-open {
-      color: green;
-    }
-  }
-  .commentLike {
-    display: flex;
-    margin: 15px 0;
-
-    p {
-      margin-left: 10px;
-      line-height: 18px;
-    }
+.commentList {
+  li {
+    font-size: 15px;
   }
 
-  .commentList {
-    li {
-      font-size: 15px;
-    }
-
-    .commentMargin {
-      margin-bottom: 5px;
-    }
-    .commentWritter {
-      font-weight: 700;
-      margin-right: 5px;
-    }
+  .commentMargin {
+    margin-bottom: 5px;
   }
-}
-
-.comment {
-  padding: 15px 20px;
-  border-top: var(--border);
-
-  input {
-    width: 400px;
-    border: none;
-    font-size: 16px;
-    outline: none;
-  }
-
-  #submit {
-    float: right;
-    border: none;
-    background-color: white;
-    font-size: 18px;
-    outline: none;
-    cursor: pointer;
+  .commentWritter {
+    font-weight: 700;
+    margin-right: 5px;
   }
 }
